Implement Filtering and Change App Icon
Use a facade style architecture for filtering,
including a FilterManger that globally manages
the current state of the filter.

This model seems to be well suited for our use case,
where multiple view controllers need to access a single
filter state.

Also, the app icon was updated to the L icon.