Implement New Auth System and Highlighted Button States
App now functions with a new auth system. Certain auth
states still need to be handled, but the large majority of
the auth flow is complete, including token caching and refreshing.

In addition, many buttons now have highlighted states, which
are created by dynamically making an image from a color.