Support fade-in animation for progressive JPEGs
Image views throughout the app now fade-in, even
with the use of progressive JPEGs. This fade-in actually
is not consistent when scrolling in a collection view and thus,
needs to be revisited in the future.

This commit also fixes an issue with the new product identifiers
being too large and breaking error checking logic. Product loading is now
much more stable.